# src/auslib/web/admin/swagger/api.yml
swagger: '2.0'
info:
  version: '1.0'
  title: "Balrog's Admin API"
  termsOfService: "https://wiki.mozilla.org/Balrog"
  license:
    name: "Mozilla Public License 2.0"
    url: "https://www.mozilla.org/en-US/MPL/2.0/"
  contact:
    name: "Mozilla Release Engineering"
    url: "https://github.com/mozilla-releng/balrog"
    email: "release-engineering@lists.mozilla.org"
  description: >
    **Balrog** is the software that runs the server side component of the update
    system used by Firefox and other Mozilla products. It is the successor to *AUS*
    (Application Update Service), which did not scale to our current needs nor allow
    us to adapt to more recent business requirements. Balrog helps us ship updates faster
    and with much more flexibility than we’ve had in the past.

# host and scheme can be omitted for a more dynamic association.
# In this case, the host and scheme used to serve the API documentation
# will be used for API calls. For example, if Swagger UI-based documentation is
# hosted at https://api.example.com/apidocs/index.html, “try it out” API calls
# will be directed to https://api.example.com.

#schemes:
#  - http
#  - https
# basePath: "/api"
# host: "localhost:8080"
# Swagger-UI bug : fix TODO week 13
...
parameters:
...
  change_id_body_param:
    name: object_change_id
    in: body
    description: change_id of the Rule/Release/Scheduled_Change to revert back to
    required: true
    schema:
      type: object
      required:
        - change_id
      properties:
        change_id:
          type: integer
          description: change_id of the Rule/Release/Scheduled_Change revision to revert to
          minimum: 0
          example: 85142
...
  changedByParam:
    name: changed_by
    in: query
    description: The username who made the requested changes.
    type: string
    format: ascii
    x-nullable: true
    required: false
  changeIdParam:
    name: change_id
    in: query
    description: The change_id value to filter hsitory by.
    type: integer
    x-nullable: true
    required: false
  historyDataVersionParam:
    name: data_version
    in: query
    description: data_version field value to filter history by.
    type: integer
    x-nullable: true
    required: false
...
  timestampFromParam:
    name: timestamp_from
    in: query
    description: timestamp from/above which to filter history.
    type: integer
    x-nullable: true
    required: false
  timestampToParam:
    name: timestamp_to
    in: query
    description: timestamp to/below which history is filtered.
    type: integer
    x-nullable: true
    required: false
  channelParam:
    name: channel
    in: query
    description: channel stirng for filtering history.
    type: string
    format: ascii
    x-nullable: true
    required: false
...
paths:
...
  /permissions/history:
    get:
      summary: Returns a list of permissions and it's scheduled change history.
      description: >
        Returns a list of permissions and it's scheduled change history.
      tags:
        - "History Permissions"
      operationId: auslib.web.common.history_all.permissions
      consumes: []
      produces: [application/json]
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#put"
        description: "Returns a list of permissions and permissions scheduled change history"
      parameters:
        - $ref: '#/parameters/productParam'
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/pageParam'
        - $ref: '#/parameters/changedByParam'
        - $ref: '#/parameters/changeIdParam'
        - $ref: '#/parameters/historyDataVersionParam'
        - $ref: '#/parameters/timestampFromParam'
        - $ref: '#/parameters/timestampToParam'
      responses:
        '200':
          description: Successfully fetched permissions and it's scheduled change history
        '400':
          $ref: '#/responses/invalidFormData'
  /required_signoffs/product/history:
    get:
      summary: Returns a list of product required signoff and it's scheduled change history.
      description: >
        Returns product required signoff and it's scheduled change history.
      tags:
        - "History Signoff Product"
      operationId: auslib.web.common.history_all.product_required_signoffs
      consumes: []
      produces: [application/json]
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#put"
        description: "Returns a list of product required signoff and it's scheduled change history"
      parameters:
        - $ref: '#/parameters/productParam'
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/pageParam'
        - $ref: '#/parameters/changedByParam'
        - $ref: '#/parameters/changeIdParam'
        - $ref: '#/parameters/historyDataVersionParam'
        - $ref: '#/parameters/timestampFromParam'
        - $ref: '#/parameters/timestampToParam'
        - $ref: '#/parameters/channelParam'
        - name: role
          in: query
          description: The role(s) required for product signoff
          type: string
          format: ascii
          x-nullable: true
          required: false
      responses:
        '200':
          description: Successfully fetched product required signoffs and it's scheduled change history
        '400':
          $ref: '#/responses/invalidFormData'
  /required_signoffs/permissions/history:
    get:
      summary: Returns a list of permissions required signoffs and it's scheduled change history.
      description: >
        Returns permissions required signoff and it's scheduled change history.
      tags:
        - "History Permissions Signoff"
      operationId: auslib.web.common.history_all.permissions_required_signoffs
      consumes: []
      produces: [application/json]
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#put"
        description: "Returns a list of permissions required signoff and it's scheduled change history"
      parameters:
        - $ref: '#/parameters/productParam'
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/pageParam'
        - $ref: '#/parameters/changedByParam'
        - $ref: '#/parameters/changeIdParam'
        - $ref: '#/parameters/historyDataVersionParam'
        - $ref: '#/parameters/timestampFromParam'
        - $ref: '#/parameters/timestampToParam'
        - name: role
          in: query
          description: The role(s) required for permission signoff
          type: string
          format: ascii
          x-nullable: true
          required: false
      responses:
        '200':
          description: Successfully fetched permissions required signoff and it's scheduled change history
        '400':
          $ref: '#/responses/invalidFormData'
...
  /users:
    get:
      summary: Returns all of Users in Balrog's DB
      description: >
        Returns all of the usernames present in Balrog. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id20)
      tags:
        - Users
      operationId: auslib.web.admin.views.permissions.get_users
      consumes: []
      produces:
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id19"
        description: "Returns all of Users in Balrog's DB."
      responses:
        '200':
          description: Get all Users
          schema:
            type: object
            properties:
              users:
                description: list of users
                type: array
                uniqueItems: true
                minItems: 0
                items:
                  type: object
                  additionalProperties:
                    type: object
                    properties:
                      permissions:
                        $ref: "#/definitions/UserPermissionsGET"
                      roles:
                        description: "user's assigned roles"
                        type: object
                        # Cannot fully validate roles object as patternProperties, part of jsonschema, isn't supported by
                        # swagger. https://github.com/OAI/OpenAPI-Specification/issues/687
                        example:
                          sample_role:
                            data_version: 1
        '401':
          $ref: '#/responses/unauthenticatedUser'
        '403':
          $ref: '#/responses/unauthorizedUser'
...
  /rules/{rule_id}/revisions:
    get:
      summary: Fetches all revisions of a Rule
      description: Fetches list and count of all the revisions of the rule given by rule_id
      tags:
        - Rules
      operationId: auslib.web.common.rules.get_history
      consumes: []
      produces:
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/database.html#rules"
        description: "Fetches list and count of all the revisions of the rule given by rule_id"
      parameters:
        - $ref: '#/parameters/rule_id_param'
        - $ref: '#/parameters/pageParam'
        - $ref: '#/parameters/limitParam'
      responses:
        '200':
          description: Returns JSON object of list of rules with their count
          schema:
            type: object
            allOf:
              - $ref: "#/definitions/RulesListGET"
              - type: object
                properties:
                  rules:
                    description: array where each element is a rule object plus timestamp, change_id and changed_by
                    type: array
                    items:
                      $ref: '#/definitions/HistoryModel'
          examples:
            application/json:
              count: 1
              rules:
                - alias: "esr45-localtest"
                  backgroundRate: 100
                  buildID: null
                  buildTarget: null
                  channel: "esr-localtest"
                  comment: "esr45 esr-localtest rule"
                  changed_by: "balrogadmin"
                  change_id: 10
                  timestamp: 149530544
                  data_version: 22
                  distVersion: null
                  distribution: null
                  fallbackMapping: null
                  headerArchitecture: null
                  instructionSet: null
                  locale: null
                  mapping: "Firefox-45.7.0esr-build1"
                  memory: null
                  mig64: null
                  osVersion: null
                  priority: 43
                  product: "Firefox"
                  rule_id: 308
                  update_type: "minor"
                  version: null
                  allowlist: null
        '404':
          $ref: '#/responses/resourceNotFound'
        '400':
          $ref: '#/responses/invalidFormData'

    post:
      summary: Reverts the rule to a previous revision
      description: >
        Updates the rule identified by the rule_id to a previous revision identified by change_id. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#history-view-object-change-id-field)
      tags:
        - Rules
      operationId: auslib.web.admin.views.rules.revert_to_older_revision
      consumes:
        - application/json
      produces:
        - application/json
        - text/html
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#history-view-object-change-id-field"
        description: "Updates the rule identified by the rule_id to a previous revision identified by change_id."
      parameters:
        - $ref: '#/parameters/rule_id_param'
        - $ref: '#/parameters/change_id_body_param'
      responses:
        '200':
          description: Successfully reverted back the Rule
        '400':
          $ref: '#/responses/invalidFormData'
        '401':
          $ref: '#/responses/unauthenticatedUser'
        '403':
          $ref: '#/responses/unauthorizedUser'
        '404':
          $ref: '#/responses/resourceNotFound'
...
definitions:
...
  OptionsModel:
    title: Options object
    description: "Data model for permission's options field"
    type: object
    properties:
      options:
        type: ['object', 'null']
        description: Options available for the given permission
        properties:
          products:
            type: array
            uniqueItems: true
            items:
              type: string
              format: ascii
            example: ["Firefox"]
          actions:
            type: array
            uniqueItems: true
            items:
              type: string
              format: ascii
            example: ["modify"]
  UserPermissionsModel:
    title: User Permissions
    description: "Model of user's permissions with all details of avaliable options and actions for the rule,release,required signoff and permission of user is assigned"
    type: object
    allOf:
      - required:
        - data_version
      - $ref: "#/definitions/DataVersionModel"
      - $ref: "#/definitions/OptionsModel"
  UserPermissionsGET:
    title: GET User Permissions
    description: Model which defines all possible user permission objects and their actions and options
    type: object
    properties:
      # Each key is a permission object
      admin:
        description: "An admin user with no options specified has completely unrestricted access to Balrog"
        allOf:
          - required:
            - data_version
          - $ref: "#/definitions/DataVersionModel"
          - type: object
            properties:
              options:
                type: ['object', 'null']
                description: Options available for admin permission
                properties:
                  products:
                    description: If specified, the user can perform any actions on Rules or Releases that affect the specified products
                    type: array
                    uniqueItems: true
                    items:
                      type: string
                      format: ascii

      release:
        description: "Release permission to perform actions on releases. Example actions: add create, delete or modify release of products"
        $ref: "#/definitions/UserPermissionsModel"

      rule:
        description: "Rule permission to perform actions. Example actions: add create, delete or modify rules of products"
        $ref: "#/definitions/UserPermissionsModel"

      required_signoff:
        description: "Permission to perform actions related to required signoffs. Example: add create, delete or modify rules of products"
        $ref: "#/definitions/UserPermissionsModel"

      release_locale:
        description: "Release Locale permission to perform modify action only on product releases."
        $ref: "#/definitions/UserPermissionsModel"

      release_read_only:
        description: "Read only Releases permission to perform modify action only on product releases."
        $ref: "#/definitions/UserPermissionsModel"

      permission:
        description: "permission to perform modify actions only."
        allOf:
          - required:
            - data_version
          - $ref: "#/definitions/DataVersionModel"
          - type: object
            properties:
              options:
                type: ['object', 'null']
                description: Options available for the given permission
                properties:
                  actions:
                    type: array
                    uniqueItems: true
                    items:
                      type: string
                      format: ascii

      scheduled_change:
        description: "permission to schedule a change to rule, release or permission. Only the Balrog Agent should be granted this permission."
        allOf:
          - required:
            - data_version
          - $ref: "#/definitions/DataVersionModel"
          - type: object
            properties:
              options:
                type: ['object', 'null']
                description: Options available for the given permission
                properties:
                  actions:
                    type: array
                    uniqueItems: true
                    items:
                      type: string
                      format: ascii

      pinnable_release:
        description: "permission to perform actions on release pins"
        $ref: "#/definitions/UserPermissionsModel"
...